requirements.txt
et_xmlfile==2.0.0
lxml==6.0.2
numpy==2.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
streamlit>=1.37
pandas>=2.1
plotly>=5.24
pyyaml>=6.0
openpyxl>=3.1

